feat(admin): Improve navigation and modal responsiveness
Introduce a scroll-to-top behavior on view changes in the AdminPage. This enhances user experience by always presenting the content at the top when navigating between different sections.

Also, refactor the OrdersView and OrderDetailModal to improve responsiveness. OrdersView now hides the list of orders when a single order is selected on larger screens, providing a dedicated space for details. The OrderDetailModal has been adapted for mobile-first design, appearing as a full-screen element on smaller devices and a centered modal on larger screens. This ensures optimal usability across various screen sizes.